double -> float

+++ Particle.h
@@ -4,7 +4,7 @@
 struct Particle
 {
     Particle() = default;
-    Particle(double mass, double position[2], double velocity[2])
+    Particle(float mass, float position[2], float velocity[2])
         : mass(mass)  
     {
         std::copy(position, position + 2, this->position);
@@ -13,7 +13,7 @@ struct Particle
     Particle(const Particle&) = default;
     Particle& operator=(const Particle&) = default;
 
-    double mass;
-    double position[2];
-    double velocity[2];
+    float mass;
+    float position[2];
+    float velocity[2];
 };
